Simulate a (Augmented) Dickey-Fuller Test to Assess Unit Root in a Time Series

sadf_test() provides a simulation approach to assessing unit root in a time series by way of the (Augmented) Dickey-Fuller test. It takes a vector and performs three (Augmented) Dickey-Fuller tests (no drift, no trend; drift, no trend; drift and trend) and calculates tau statistics as one normally would. Rather than interpolate or approximate a p-value, it simulates some user-specified number of (Augmented) Dickey-Fuller tests of either a known, non-stationary time series or a known, white-noise time series matching the length of the time series the user provides. This allows the user to make assessments of non-stationarity or stationarity by way of simulation rather than approximation from received critical values by way of books or tables some years out of date.

Usage

sadf_test(x, n_lags = NULL, n_sims = 1000, sim_hyp = "nonstationary")

Arguments

x

a vector

n_lags

defaults to NULL, but must be 0 or a positive integer. This argument determines the number of lagged first differences to include in the estimation procedure. Recall that the test statistic (tau) is still the t-statistic for the level value of the vector at t-1, whether the constant (drift) and time trend is included or not. If this value is 0, the procedure is the classic Dickey-Fuller test. If this value is greater than 0, this is the "augmented" Dickey-Fuller test, so-called because it is "augmented" by the number of lagged first differences to assess higher-order AR processes. If no argument is specified, the default lag is the length of the series, divided over 100, raised to the power of 1/4, and floored. Check "Details" and "Examples" for more information.

n_sims

the number of simulations for calculating an interval or distribution of test statistics for assessing stationarity or non-stationarity. Defaults to 1,000.

sim_hyp

can be either "stationary" or "nonstationary". If "stationary", the function runs (A)DF tests on simulated stationary (pure white noise) data. This allows the user to assess compatibility/plausibility of the test statistic against a distribution of test statistics that are known to be pure white noise (in expectation). If "nonstationary" (default), the function generates three different data sets of a pure random walk, a random walk with a drift, and a random walk with a drift and trend. It then runs (A)DF tests on all those. This allows the user to assess the compatibility/plausibility of their test statistics with data that are known to be nonstationary in some form.

Value

sadf_test() returns a list of length 3. The first element in the list is a matrix of tau statistics calculated by the test. The second element is a data frame of the simulated tau statistics of either a known white-noise time series or three different non-stationary time series (pure random walk, random walk with drift, random walk with drift and trend). The third element contains some attributes about the procedure for post-processing.

Details

The Dickey-Fuller and its "augmented" corollary are curious statistical procedures, even if the underlying concept is straightforward. I have seen various implementations of these procedures use slightly different terminology to describe its procedure, though this particular implementation will impose nomenclature in which the classic Dickey-Fuller procedure that assumes just the AR(1) process is one in which n_lags is 0. The addition of lags (of first differences) is what ultimately makes the Dickey-Fuller procedure to be "augmented."

The function employs the default suggested by Schwert (1989) for the number of lagged first differences to include in this procedure. Schwert (1989) recommends taking the length of the series and dividing it by 100 before raising that number to the power of 1/4. Thereafter, multiply it by 12 and round down the number to the nearest integer. There are other suggested defaults you can consider. adf.test in aTSA takes the length of the series, divides it by 100 and raises it to the power of 2/9. It multiplies that by 4 and floors the result. adf.test in tseries subtracts 1 from the length of the series before raising it to the power of 1/3 (flooring that result as well). The Examples section will show you how you can do this.

This function specifies three different types of tests: 1) no drift, no trend, 2) drift, no trend, and 3) drift and trend. In the language of the lm() function, the first is lm(y ~ ly - 1) where y is the value of y and ly is its first-order lag. The second test is lm(y ~ ly), intuitively suggesting the y-intercept in this equation is the "drift". The third would be lm(y ~ ly + t) with t being a simple integer that increases by 1 for each observation (i.e. a time-trend).

None of this is meant to discourage the use of Fuller (1976) or its various reproductions for the sake of diagnosing stationarity or non-stationary, and I will confess their expertise on these matters outpaces mine. Consider the justification for this function to be largely philosophical and/or experimental. Why not simulate it? It's not like time or computing power are huge issues anymore.

This is always awkwardly stated, but it's a good reminder that the classic Dickey-Fuller statistics are mostly intended to come back negative. That's not always the case, to be clear, but it is the intended case. You assess the statistic by "how negative" it is. Stationary time series will produce test statistics more negative ("smaller") than those produced by non-stationary time series. In a way, this makes the hypotheses implicitly one-tailed (to use that language).

This function removes missing values from the vector before calculating test statistics.

References

Schwert, G. William. 1989. "Tests for Unit Roots: A Monte Carlo Investigation." Journal of Business & Economic Statistics 7(2): 147–159.

Author

Steven V. Miller

Examples

y <- na.omit(USDSEK[1:500,])$close

# Default suggested number of lags, by way of Schwert (1989):
floor(12*(length(y)/100)^(.25))
#> [1] 17
# Default method in {tseries}' adf.test()
floor((length(y)-1)^(1/3)) # Said and Dickey (1984)
#> [1] 7
# Default method for lags in {aTSA}' adf.test()
floor(4*(length(y)/100)^(2/9))
#> [1] 5

sadf_test(y, n_sims = 25) # Doing 25, just to make it quick
#> $stats
#>            [,1]
#> [1,] -0.4805752
#> [2,] -1.5710219
#> [3,] -1.8676060
#> 
#> $sims
#>            tau sim                cat
#> 1  -0.23052716   1 No Drift, No Trend
#> 2  -1.66264661   1    Drift, No Trend
#> 3  -3.11928098   1    Drift and Trend
#> 4  -0.23146614   2 No Drift, No Trend
#> 5  -2.25284709   2    Drift, No Trend
#> 6  -2.36376986   2    Drift and Trend
#> 7   0.62464928   3 No Drift, No Trend
#> 8  -0.35811748   3    Drift, No Trend
#> 9  -1.87199271   3    Drift and Trend
#> 10 -0.89493102   4 No Drift, No Trend
#> 11 -1.11942151   4    Drift, No Trend
#> 12 -3.18553732   4    Drift and Trend
#> 13 -2.39915745   5 No Drift, No Trend
#> 14 -2.59456641   5    Drift, No Trend
#> 15 -1.44317099   5    Drift and Trend
#> 16 -1.51952880   6 No Drift, No Trend
#> 17 -1.72160338   6    Drift, No Trend
#> 18 -2.79921703   6    Drift and Trend
#> 19  0.73296651   7 No Drift, No Trend
#> 20 -0.26369900   7    Drift, No Trend
#> 21 -2.50577133   7    Drift and Trend
#> 22 -0.83653426   8 No Drift, No Trend
#> 23 -0.83266594   8    Drift, No Trend
#> 24 -2.53569412   8    Drift and Trend
#> 25  0.08986507   9 No Drift, No Trend
#> 26 -2.43344412   9    Drift, No Trend
#> 27 -2.29322889   9    Drift and Trend
#> 28  0.42576518  10 No Drift, No Trend
#> 29 -1.91164701  10    Drift, No Trend
#> 30 -1.93934200  10    Drift and Trend
#> 31 -1.23789372  11 No Drift, No Trend
#> 32 -1.23748972  11    Drift, No Trend
#> 33 -3.18697808  11    Drift and Trend
#> 34 -0.61827235  12 No Drift, No Trend
#> 35 -1.82665714  12    Drift, No Trend
#> 36 -3.59242172  12    Drift and Trend
#> 37 -0.29863890  13 No Drift, No Trend
#> 38 -1.36419849  13    Drift, No Trend
#> 39 -2.11031425  13    Drift and Trend
#> 40  0.52939735  14 No Drift, No Trend
#> 41 -1.12430021  14    Drift, No Trend
#> 42 -2.70606958  14    Drift and Trend
#> 43 -1.16783534  15 No Drift, No Trend
#> 44 -2.69048958  15    Drift, No Trend
#> 45 -1.77509245  15    Drift and Trend
#> 46 -1.22448805  16 No Drift, No Trend
#> 47 -1.15900616  16    Drift, No Trend
#> 48 -2.20946726  16    Drift and Trend
#> 49 -0.09785916  17 No Drift, No Trend
#> 50 -1.07974189  17    Drift, No Trend
#> 51 -3.69890355  17    Drift and Trend
#> 52  1.51302455  18 No Drift, No Trend
#> 53  0.91548561  18    Drift, No Trend
#> 54 -3.79044292  18    Drift and Trend
#> 55 -0.33995095  19 No Drift, No Trend
#> 56 -1.81012114  19    Drift, No Trend
#> 57 -2.62109623  19    Drift and Trend
#> 58  2.43426508  20 No Drift, No Trend
#> 59 -0.71173159  20    Drift, No Trend
#> 60 -2.15473402  20    Drift and Trend
#> 61 -0.55832595  21 No Drift, No Trend
#> 62 -2.91380432  21    Drift, No Trend
#> 63 -2.91014320  21    Drift and Trend
#> 64 -0.74425523  22 No Drift, No Trend
#> 65 -2.02322764  22    Drift, No Trend
#> 66 -2.24007677  22    Drift and Trend
#> 67 -1.02013438  23 No Drift, No Trend
#> 68 -0.91705959  23    Drift, No Trend
#> 69 -1.96955918  23    Drift and Trend
#> 70  0.25182829  24 No Drift, No Trend
#> 71 -1.99257048  24    Drift, No Trend
#> 72 -2.12363011  24    Drift and Trend
#> 73 -0.07646436  25 No Drift, No Trend
#> 74 -0.10393846  25    Drift, No Trend
#> 75 -1.61942907  25    Drift and Trend
#> 
#> $attributes
#>   lags       sim_hyp n_sims   n test
#> 1   17 nonstationary     25 499  adf
#> 
#> attr(,"class")
#> [1] "sadf_test"