Simulate a Phillips-Perron Test to Assess Unit Root in a Time Series

spp_test() provides a simulation approach to assessing unit root in a time series by way of the Phillips-Perron test. It takes a vector and performs three Phillips-Perron tests (no drift, no trend; drift, no trend; drift and trend) and calculates both rho and tau statistics as one normally would. Rather than interpolate or approximate a p-value, it simulates some user-specified number of Phillips-Perron tests of either a known, non-stationary time series or a known, white-noise time series matching the length of the time series the user provides. This allows the user to make assessments of non-stationarity or stationarity by way of simulation rather than approximation from received critical values by way of books or tables some years out of date.

Usage

spp_test(x, lag_short = TRUE, n_sims = 1000, sim_hyp = "nonstationary")

Arguments

x

a vector

lag_short

logical, defaults to TRUE. If TRUE, the "short-term" lag is used for the Phillips-Perron test. If FALSE, the "long-term" lag is used.

n_sims

the number of simulations for calculating an interval or distribution of test statistics for assessing stationarity or non-stationarity. Defaults to 1,000.

sim_hyp

can be either "stationary" or "nonstationary". If "stationary", the function runs Phillips-Perron tests on simulated stationary (pure white noise) data. This allows the user to assess compatibility/plausibility of the test statistic against a distribution of test statistics that are known to be pure white noise (in expectation). If "nonstationary" (default), the function generates three different data sets of a pure random walk, a random walk with a drift, and a random walk with a drift and trend. It then runs Phillips-Perron tests on all those. This allows the user to assess the compatibility/plausibility of their test statistics with data that are known to be nonstationary in some form.

Value

spp_test() returns a list of length 3. The first element in the list is a matrix of rho statistics and tau statistics calculated by the Phillips-Perron test. The second element is a data frame of the simulated rho and tau statistics of either a known white-noise time series or three different non-stationary time series (pure random walk, random walk with drift, random walk with drift and trend). The third element is some attributes about the procedure for post-processing.

Details

Some knowledge of Augmented Dickey-Fuller and the Phillips-Perron procedure is assumed here. Generally, the Phillips-Perron test purports to build on the Augmented Dickey-Fuller procedure through two primary means. The first is relaxing the need to specify or assume lag structures ad hoc or ex ante. Only a short-term lag or long-term lag are necessary. The second is that its robust to various forms of heteroskedasticity in the error term.

The short-term and long-term lags follow the convention introduced in the Phillips-Perron test. The short-term lag uses the default number of Newey-West lags, defined as the floor of 4*(n/100)^.25 where n is the length of the time series. The long-term lag substitutes 4 for 12 in this equation. We have Schwert (1989) to thank for these defaults.

This function specifies three different types of tests: 1) no drift, no trend, 2) drift, no trend, and 3) drift and trend. In the language of the lm() function, the first is lm(y ~ ly - 1) where y is the value of y and ly is its first-order lag. The second test is lm(y ~ ly), intuitively suggesting the y-intercept in this equation is the "drift". The third would be lm(y ~ ly + t) with t being a simple integer that increases by 1 for each observation (i.e. a time-trend).

There are two types of statistics in the Phillips-Perron test: rho and tau. Of the two, tau is the more intuitive statistic and compares favorably to its corollary statistic in the Augmented Dickey-Fuller test. It's why you'll typically see tau reported as the statistic of interest in other implementations. rho has its utility for more advanced diagnostics, though. Both are calculated in this function, though tau is the default statistic.

None of this is meant to discourage the use of Fuller (1976) or its various reproductions for the sake of diagnosing stationarity or non-stationary, and I will confess their expertise on these matters outpaces mine. Consider the justification for this function to be largely philosophical and/or experimental. Why not simulate it? It's not like time or computing power are huge issues anymore.

This is always awkwardly stated, but it's a good reminder that the classic Dickey-Fuller statistics are mostly intended to come back negative. That's not always the case, to be clear, but it is the intended case. You assess the statistic by "how negative" it is. Stationary time series will produce test statistics more negative ("smaller") than those produced by non-stationary time series. In a way, this makes the hypotheses implicitly one-tailed (to use that language).

This function removes missing values from the vector before calculating test statistics.

Author

Steven V. Miller

Examples

a <- rnorm(25) # white noise
b <- cumsum(a) # random walk

spp_test(a, n_sims = 25)
#> $stats
#>           [,1]      [,2]
#> [1,] -27.13585 -5.936353
#> [2,] -27.12847 -6.100783
#> [3,] -26.83813 -6.054596
#> 
#> $sims
#>           z_rho       z_tau sim                cat
#> 1   -2.82861295 -1.22214393   1 No Drift, No Trend
#> 2   -6.45295200 -1.98190002   1    Drift, No Trend
#> 3  -15.67503152 -3.95472279   1    Drift and Trend
#> 4   -5.21221813 -1.62458532   2 No Drift, No Trend
#> 5   -5.08748670 -1.54378739   2    Drift, No Trend
#> 6   -2.78371754 -1.00326730   2    Drift and Trend
#> 7    1.13483923  1.99997739   3 No Drift, No Trend
#> 8   -0.46914230 -0.53837696   3    Drift, No Trend
#> 9  -11.30257013 -2.46371591   3    Drift and Trend
#> 10   0.48733892  0.38061494   4 No Drift, No Trend
#> 11  -1.61666512 -0.88654502   4    Drift, No Trend
#> 12 -11.94953589 -2.77783733   4    Drift and Trend
#> 13  -0.52177555 -0.40088708   5 No Drift, No Trend
#> 14  -8.73230063 -3.43721159   5    Drift, No Trend
#> 15  -3.23930206 -1.27895875   5    Drift and Trend
#> 16  -5.03962964 -1.49498542   6 No Drift, No Trend
#> 17  -4.95612618 -1.47226191   6    Drift, No Trend
#> 18  -4.81933941 -1.12715319   6    Drift and Trend
#> 19  -1.97464138 -1.00780012   7 No Drift, No Trend
#> 20  -7.71435158 -2.02535164   7    Drift, No Trend
#> 21 -17.55785589 -4.36653913   7    Drift and Trend
#> 22 -15.02935495 -3.32645752   8 No Drift, No Trend
#> 23 -15.62685653 -3.46961159   8    Drift, No Trend
#> 24  -4.65695654 -1.61338366   8    Drift and Trend
#> 25  -0.58906951 -0.49357437   9 No Drift, No Trend
#> 26  -8.27413109 -2.96987298   9    Drift, No Trend
#> 27 -13.03811245 -3.06419187   9    Drift and Trend
#> 28   0.05755162  0.05798358  10 No Drift, No Trend
#> 29  -4.04690257 -2.17465819  10    Drift, No Trend
#> 30  -7.09034108 -2.25110676  10    Drift and Trend
#> 31   1.37808220  1.36710495  11 No Drift, No Trend
#> 32  -2.14647854 -1.14752972  11    Drift, No Trend
#> 33  -6.43153145 -1.85968240  11    Drift and Trend
#> 34  -0.73091816 -0.49096525  12 No Drift, No Trend
#> 35  -4.02791809 -1.80104617  12    Drift, No Trend
#> 36  -4.59011030 -1.14879814  12    Drift and Trend
#> 37  -1.77374508 -0.68310727  13 No Drift, No Trend
#> 38  -2.50393754 -0.88471118  13    Drift, No Trend
#> 39  -3.26668162 -0.89136201  13    Drift and Trend
#> 40  -1.84259666 -0.79603756  14 No Drift, No Trend
#> 41 -11.66058884 -2.55425083  14    Drift, No Trend
#> 42  -9.19982564 -2.20010207  14    Drift and Trend
#> 43   0.96355749  2.00989913  15 No Drift, No Trend
#> 44  -1.37087319 -1.80430395  15    Drift, No Trend
#> 45  -3.80809321 -1.12700450  15    Drift and Trend
#> 46  -1.05420567 -0.48232439  16 No Drift, No Trend
#> 47  -5.50646924 -1.82859960  16    Drift, No Trend
#> 48  -7.79745006 -2.90252078  16    Drift and Trend
#> 49  -1.39575002 -0.71855444  17 No Drift, No Trend
#> 50  -0.80981386 -0.38274115  17    Drift, No Trend
#> 51  -7.54223664 -1.95422034  17    Drift and Trend
#> 52  -0.44713603 -0.37545526  18 No Drift, No Trend
#> 53  -2.86314767 -1.50815366  18    Drift, No Trend
#> 54 -10.30611161 -2.33257307  18    Drift and Trend
#> 55  -9.84169660 -2.17419621  19 No Drift, No Trend
#> 56 -10.85498937 -2.30882376  19    Drift, No Trend
#> 57 -15.12517176 -3.17027211  19    Drift and Trend
#> 58  -2.45742794 -0.89125508  20 No Drift, No Trend
#> 59  -2.30547251 -0.85412930  20    Drift, No Trend
#> 60 -10.61146915 -2.38367760  20    Drift and Trend
#> 61  -2.58777164 -1.12687636  21 No Drift, No Trend
#> 62  -2.96324042 -1.14308156  21    Drift, No Trend
#> 63  -2.63735385 -0.89461718  21    Drift and Trend
#> 64  -8.34807331 -2.06234904  22 No Drift, No Trend
#> 65  -9.56487881 -2.20917590  22    Drift, No Trend
#> 66  -6.64817372 -1.56047562  22    Drift and Trend
#> 67  -1.19314692 -0.75234412  23 No Drift, No Trend
#> 68 -13.38406552 -3.81773692  23    Drift, No Trend
#> 69  -8.46502171 -2.17955084  23    Drift and Trend
#> 70   0.49881236  0.70715720  24 No Drift, No Trend
#> 71  -2.16851826 -1.75545110  24    Drift, No Trend
#> 72  -4.54435817 -1.59360463  24    Drift and Trend
#> 73  -1.17285430 -0.64597235  25 No Drift, No Trend
#> 74 -11.14177246 -3.32721355  25    Drift, No Trend
#> 75 -10.63436580 -2.31647669  25    Drift and Trend
#> 
#> $attributes
#>   lags       sim_hyp n_sims  n test
#> 1    2 nonstationary     25 25   pp
#> 
#> attr(,"class")
#> [1] "spp_test"
spp_test(b, n_sims = 25)
#> $stats
#>            [,1]       [,2]
#> [1,]   1.173118  0.7177936
#> [2,]  -2.699879 -0.7033299
#> [3,] -15.058437 -2.8693332
#> 
#> $sims
#>           z_rho       z_tau sim                cat
#> 1    2.44006356  2.28389767   1 No Drift, No Trend
#> 2    1.34762917  0.88848299   1    Drift, No Trend
#> 3   -6.49278202 -1.77518639   1    Drift and Trend
#> 4    0.16396016  0.18478655   2 No Drift, No Trend
#> 5   -2.83884146 -2.16295526   2    Drift, No Trend
#> 6   -5.88496580 -1.84019319   2    Drift and Trend
#> 7   -0.49949351 -0.31196900   3 No Drift, No Trend
#> 8  -10.73589622 -2.88833524   3    Drift, No Trend
#> 9   -6.15325781 -2.15052882   3    Drift and Trend
#> 10   1.60816348  2.11225920   4 No Drift, No Trend
#> 11  -0.49690254 -0.34096563   4    Drift, No Trend
#> 12  -2.82781117 -1.11737569   4    Drift and Trend
#> 13  -3.68977592 -1.44760002   5 No Drift, No Trend
#> 14  -4.24083001 -1.57813971   5    Drift, No Trend
#> 15  -4.41887270 -1.11370688   5    Drift and Trend
#> 16   0.33055758  0.32027431   6 No Drift, No Trend
#> 17  -1.30883217 -0.98553869   6    Drift, No Trend
#> 18  -9.45426878 -2.35173155   6    Drift and Trend
#> 19  -3.27365064 -1.45788345   7 No Drift, No Trend
#> 20  -4.45722422 -2.36444257   7    Drift, No Trend
#> 21  -5.72415840 -1.32820504   7    Drift and Trend
#> 22  -7.40443469 -1.96168172   8 No Drift, No Trend
#> 23 -13.99779197 -3.47927307   8    Drift, No Trend
#> 24  -9.99155825 -2.30813281   8    Drift and Trend
#> 25  -2.28602384 -0.93402196   9 No Drift, No Trend
#> 26  -1.88524192 -0.63516702   9    Drift, No Trend
#> 27  -5.23815526 -1.51729509   9    Drift and Trend
#> 28   1.41567456  2.79629063  10 No Drift, No Trend
#> 29   0.07144525  0.09505365  10    Drift, No Trend
#> 30 -11.35470139 -2.47758872  10    Drift and Trend
#> 31 -13.97339847 -2.85441976  11 No Drift, No Trend
#> 32 -14.37457823 -2.88565023  11    Drift, No Trend
#> 33  -7.49576482 -1.83089209  11    Drift and Trend
#> 34  -1.33088527 -0.80551949  12 No Drift, No Trend
#> 35  -3.66819849 -1.22047917  12    Drift, No Trend
#> 36  -7.95924493 -1.51367519  12    Drift and Trend
#> 37 -13.67750749 -2.80072709  13 No Drift, No Trend
#> 38 -16.54642042 -3.19010087  13    Drift, No Trend
#> 39 -10.41479792 -2.44111310  13    Drift and Trend
#> 40  -1.23874411 -0.47449491  14 No Drift, No Trend
#> 41  -0.43284665 -0.18032917  14    Drift, No Trend
#> 42   1.12989269  0.62431221  14    Drift and Trend
#> 43  -1.04181368 -0.72212623  15 No Drift, No Trend
#> 44  -3.80974135 -1.26646579  15    Drift, No Trend
#> 45  -8.73947901 -2.19211565  15    Drift and Trend
#> 46  -1.43920999 -0.78504772  16 No Drift, No Trend
#> 47  -5.42132802 -1.75094905  16    Drift, No Trend
#> 48  -9.14306933 -2.26064491  16    Drift and Trend
#> 49  -1.23476071 -0.54958827  17 No Drift, No Trend
#> 50  -6.16108506 -1.78698235  17    Drift, No Trend
#> 51  -2.01841191 -0.62930270  17    Drift and Trend
#> 52  -4.49115693 -1.49525403  18 No Drift, No Trend
#> 53  -6.81673038 -1.81792611  18    Drift, No Trend
#> 54 -11.40449483 -2.45004204  18    Drift and Trend
#> 55   2.29216819  1.96142630  19 No Drift, No Trend
#> 56   1.58898010  1.32389376  19    Drift, No Trend
#> 57  -6.40688151 -1.58735368  19    Drift and Trend
#> 58  -1.38248988 -0.70167176  20 No Drift, No Trend
#> 59  -1.02150001 -0.54890663  20    Drift, No Trend
#> 60  -5.76522119 -1.84887885  20    Drift and Trend
#> 61  -4.36400701 -1.50577777  21 No Drift, No Trend
#> 62  -4.38948547 -1.47624561  21    Drift, No Trend
#> 63 -12.17341852 -2.71390303  21    Drift and Trend
#> 64   0.76276087  0.93232255  22 No Drift, No Trend
#> 65  -1.15672496 -0.92456104  22    Drift, No Trend
#> 66  -6.17316071 -2.36454980  22    Drift and Trend
#> 67  -0.12489080 -0.06503457  23 No Drift, No Trend
#> 68  -3.89471980 -1.26206394  23    Drift, No Trend
#> 69 -11.57172041 -2.63352595  23    Drift and Trend
#> 70 -11.59989128 -3.02681359  24 No Drift, No Trend
#> 71 -11.96025630 -2.95678198  24    Drift, No Trend
#> 72  -7.69361347 -1.91596353  24    Drift and Trend
#> 73   0.02809885  0.03949833  25 No Drift, No Trend
#> 74  -5.53114060 -2.85145643  25    Drift, No Trend
#> 75  -7.01937874 -1.78909867  25    Drift and Trend
#> 
#> $attributes
#>   lags       sim_hyp n_sims  n test
#> 1    2 nonstationary     25 25   pp
#> 
#> attr(,"class")
#> [1] "spp_test"