Daily maturity rates for U.S. Treasury Bills

A data frame on daily (when applicable/available) U.S. Treasury Bill rates. These are the yield received for investing in a government-issued treasury security that has a maturity of a given period of time (three months, six months, or a year).

Usage

tbills

Format

A data frame with 17,741 observations on the following 4 variables.

date

a date

tb3m

the three-month treasury bill rate

tb6m

the six-month treasury bill rate

tb1y

the one-year treasury bill rate

Details

Data come by way of fredr call. The one-year (52-week) treasury bill rate was discontinued in 2001 and re-introduced in 2008. Be mindful of that gap in the series.