Summarize Unit Root Test Simulations

ur_summary() provides a summary of the unit root tests included in this package.

Usage

ur_summary(obj, pp_stat = "tau", ...)

Arguments

obj

the object to be summarized, of class 'spp_test'

pp_stat

a statistic to be summarized: either "tau" or "rho". Applicable only to Phillips-Perron tests generated by functions in this package.

...

additional argument, currently ignored

Value

ur_summary() produces console output that offers a summary assessment about the presence of a unit root based on your simulations.

Details

This function makes ample use of the "attributes" element in the list produced by the unit root simulations.

Author

Steven V. Miller

Examples

A <- spp_test(money_demand$ffer, n_sims = 100)
ur_summary(A)
#> ------------------------------------------ 
#> * Simulated Phillips-Perron Test Summary * 
#> ------------------------------------------ 
#> Simulated test statistics are calculated on time series that are: nonstationary 
#> Length of time series: 259. Lags: 5
#> 
#> Type 1: no drift, no trend 
#> --------------------------
#> Your tau: -1.365
#> Potential thresholds for your consideration: -2.664 (1%); -2.065 (5%); -1.732 (10%)
#> 
#> Type 2: drift, no trend
#> -----------------------
#> Your tau: -2.391
#> Potential thresholds for your consideration: -3.042 (1%); -2.84 (5%); -2.558 (10%)
#> 
#> Type 3: drift and trend
#> -----------------------
#> Your tau: -2.771
#> Potential thresholds for your consideration: -3.774 (1%); -3.617 (5%); -3.174 (10%)
#> 
#> 
#> --------------------------------------------------------------
#> * Guides to help you assess stationarity or non-stationarity * 
#> --------------------------------------------------------------
#> These thresholds are the results of 100 different simulations of a non-stationary time series matching your time series description (n = 259, lags = 5). If your tau is more negative than one of these thresholds of interest, that is incompatible with a non-stationary time series and more compatible with a stationary time series.
#> 
#> If this is not the case, what you see is implying your time series is non-stationary.
#> 
#> Please refer to the raw output for the simulations for other means of assessment/summary.