Simulate a (Augmented) Dickey-Fuller Test to Assess Unit Root in a Time Series

sadf_test() provides a simulation approach to assessing unit root in a time series by way of the (Augmented) Dickey-Fuller test. It takes a vector and performs three (Augmented) Dickey-Fuller tests (no drift, no trend; drift, no trend; drift and trend) and calculates both rho and tau statistics as one normally would. Rather than interpolate or approximate a p-value, it simulates some user-specified number of (Augmented) Dickey-Fuller tests of a known, white-noise time series matching the length of the time series the user provides. This allows the user to make assessments of non-stationarity or stationarity by way of simulation rather than approximation from received critical values by way of books or tables some years out of date.

Usage

sadf_test(x, n_lags = NULL, n_sims = 1000, sim_hyp = "nonstationary")

Arguments

x

a vector

n_lags

defaults to NULL, but must be 0 or a positive integer. This argument determines the number of lagged first differences to include in the estimation procedure. Recall that the test statistic (tau) is still the t-statistic for the level value of the vector at t-1, whether the constant (drift) and time trend is included or not. If this value is 0, the procedure is the classic Dickey-Fuller test. If this value is greater than 0, this is the "augmented" Dickey-Fuller test, so-called because it is "augmented" by the number of lagged first differences to assess higher-order AR processes. If no argument is specified, the default lag is the length of the series, divided over 100, raised to the power of 1/4, and floored. Check "Details" and "Examples" for more information.

n_sims

the number of simulations for calculating an interval or distribution of test statistics for assessing stationarity or non-stationarity. Defaults to 1,000.

sim_hyp

can be either "stationary" or "nonstationary". If "stationary", the function runs (A)DF tests on simulated stationary (pure white noise) data. This allows the user to assess compatibility/plausibility of the test statistic against a distribution of test statistics that are known to be pure white noise (in expectation). If "nonstationary" (default), the function generates three different data sets of a pure random walk, a random walk with a drift, and a random walk with a drift and trend. It then runs (A)DF tests on all those. This allows the user to assess the compatibility/plausibility of their test statistics with data that are known to be nonstationary in some form.

Value

sadf_test() returns a list of length 3. The first element in the list is a matrix of tau statistics calculated by the test. The second element is a data frame of the simulated tau statistics of either a known white-noise time series or three different non-stationary time series (pure random walk, random walk with drift, random walk with drift and trend). The third element contains some attributes about the procedure for post-processing.

Details

The Dickey-Fuller and its "augmented" corollary are curious statistical procedures, even if the underlying concept is straightforward. I have seen various implementations of these procedures use slightly different terminology to describe its procedure, though this particular implementation will impose nomenclature in which the classic Dickey-Fuller procedure that assumes just the AR(1) process is one in which n_lags is 0. The addition of lags (of first differences) is what ultimately makes the Dickey-Fuller procedure to be "augmented."

The function employs the default suggested by Schwert (1989) for the number of lagged first differences to include in this procedure. Schwert (1989) recommends taking the length of the series and dividing it by 100 before raising that number to the power of 1/4. Thereafter, multiply it by 12 and round down the number to the nearest integer. There are other suggested defaults you can consider. adf.test in aTSA takes the length of the series, divides it by 100 and raises it to the power of 2/9. It multiplies that by 4 and floors the result. adf.test in tseries subtracts 1 from the length of the series before raising it to the power of 1/3 (flooring that result as well). The Examples section will show you how you can do this.

This function specifies three different types of tests: 1) no drift, no trend, 2) drift, no trend, and 3) drift and trend. In the language of the lm() function, the first is lm(y ~ ly - 1) where y is the value of y and ly is its first-order lag. The second test is lm(y ~ ly), intuitively suggesting the y-intercept in this equation is the "drift". The third would be lm(y ~ ly + t) with t being a simple integer that increases by 1 for each observation (i.e. a time-trend).

None of this is meant to discourage the use of Fuller (1976) or its various reproductions for the sake of diagnosing stationarity or non-stationary, and I will confess their expertise on these matters outpaces mine. Consider the justification for this function to be largely philosophical and/or experimental. Why not simulate it? It's not like time or computing power are huge issues anymore.

This is always awkwardly stated, but it's a good reminder that the classic Dickey-Fuller statistics are mostly intended to come back negative. That's not always the case, to be clear, but it is the intended case. You assess the statistic by "how negative" it is. Stationary time series will produce test statistics more negative ("smaller") than those produced by non-stationary time series. In a way, this makes the hypotheses implicitly one-tailed (to use that language).

This function removes missing values from the vector before calculating test statistics.

References

Schwert, G. William. 1989. "Tests for Unit Roots: A Monte Carlo Investigation." Journal of Business & Economic Statistics 7(2): 147–159.

Author

Steven V. Miller

Examples

a <- rnorm(25) # white noise
b <- cumsum(a) # random walk

x <- c(4, 1, 5, 8, 3, 9, 3, 6, 8, 6, 4, 4, 1, 10, 7, 6, 9, 2, 10, 9, 9, 4, 8, 3, 9)
y <- na.omit(USDSEK[1:500,])

# Default suggested number of lags, by way of Schwert (1989):
floor(12*(length(y)/100)^(.25))
#> [1] 4
# Default method in {tseries}' adf.test()
floor((length(y)-1)^(1/3)) # Said and Dickey (1984)
#> [1] 1
# Default method for lags in {aTSA}' adf.test()
floor(4*(length(y)/100)^(2/9))
#> [1] 1

sadf_test(a, n_sims = 25)
#> $stats
#>           [,1]
#> [1,] -1.256376
#> [2,] -1.132289
#> [3,] -1.018200
#> 
#> $sims
#>            tau sim                cat
#> 1  -0.57363689   1 No Drift, No Trend
#> 2  -1.70552955   1    Drift, No Trend
#> 3  -2.01233675   1    Drift and Trend
#> 4   0.35642392   2 No Drift, No Trend
#> 5  -0.84102047   2    Drift, No Trend
#> 6  -6.77751952   2    Drift and Trend
#> 7  -1.30728169   3 No Drift, No Trend
#> 8  -1.67531756   3    Drift, No Trend
#> 9  -2.56332562   3    Drift and Trend
#> 10 -0.28657238   4 No Drift, No Trend
#> 11 -0.64255940   4    Drift, No Trend
#> 12 -3.54179791   4    Drift and Trend
#> 13  1.17064213   5 No Drift, No Trend
#> 14  1.09666729   5    Drift, No Trend
#> 15 -3.45420450   5    Drift and Trend
#> 16 -0.14816735   6 No Drift, No Trend
#> 17  0.30659879   6    Drift, No Trend
#> 18 -1.46188663   6    Drift and Trend
#> 19  0.03560630   7 No Drift, No Trend
#> 20 -1.46866060   7    Drift, No Trend
#> 21 -1.01774286   7    Drift and Trend
#> 22  0.79757595   8 No Drift, No Trend
#> 23 -0.44081666   8    Drift, No Trend
#> 24 -2.52099337   8    Drift and Trend
#> 25 -1.02329596   9 No Drift, No Trend
#> 26 -2.74331476   9    Drift, No Trend
#> 27 -1.44063038   9    Drift and Trend
#> 28 -0.14033966  10 No Drift, No Trend
#> 29 -1.91314798  10    Drift, No Trend
#> 30 -0.40470309  10    Drift and Trend
#> 31 -0.78663931  11 No Drift, No Trend
#> 32 -1.07478564  11    Drift, No Trend
#> 33 -1.93427096  11    Drift and Trend
#> 34  0.83310092  12 No Drift, No Trend
#> 35  0.21755571  12    Drift, No Trend
#> 36  0.56901380  12    Drift and Trend
#> 37 -0.55118949  13 No Drift, No Trend
#> 38 -1.82172237  13    Drift, No Trend
#> 39 -0.37979629  13    Drift and Trend
#> 40 -0.43683934  14 No Drift, No Trend
#> 41 -0.72830595  14    Drift, No Trend
#> 42 -2.02475121  14    Drift and Trend
#> 43 -2.05475501  15 No Drift, No Trend
#> 44 -3.83862209  15    Drift, No Trend
#> 45 -1.58151515  15    Drift and Trend
#> 46 -1.17216089  16 No Drift, No Trend
#> 47 -0.31615360  16    Drift, No Trend
#> 48  0.91317596  16    Drift and Trend
#> 49 -0.56425572  17 No Drift, No Trend
#> 50 -0.82921301  17    Drift, No Trend
#> 51  0.51887341  17    Drift and Trend
#> 52 -0.53084601  18 No Drift, No Trend
#> 53 -1.75645306  18    Drift, No Trend
#> 54 -1.33125976  18    Drift and Trend
#> 55 -0.24116948  19 No Drift, No Trend
#> 56 -2.80830216  19    Drift, No Trend
#> 57 -2.69569654  19    Drift and Trend
#> 58  1.30924484  20 No Drift, No Trend
#> 59 -0.16687121  20    Drift, No Trend
#> 60 -0.72376989  20    Drift and Trend
#> 61 -0.12231423  21 No Drift, No Trend
#> 62 -0.95875103  21    Drift, No Trend
#> 63 -1.19308168  21    Drift and Trend
#> 64 -0.70424459  22 No Drift, No Trend
#> 65 -2.50604106  22    Drift, No Trend
#> 66 -1.30094402  22    Drift and Trend
#> 67  0.69276643  23 No Drift, No Trend
#> 68 -0.29422980  23    Drift, No Trend
#> 69 -2.47547709  23    Drift and Trend
#> 70 -0.55092735  24 No Drift, No Trend
#> 71 -2.38961953  24    Drift, No Trend
#> 72 -3.17988668  24    Drift and Trend
#> 73 -0.08383899  25 No Drift, No Trend
#> 74 -1.30630665  25    Drift, No Trend
#> 75 -3.37065270  25    Drift and Trend
#> 
#> $attributes
#>   lags       sim_hyp n_sims  n test
#> 1    8 nonstationary     25 25  adf
#> 
#> attr(,"class")
#> [1] "sadf_test"
sadf_test(b, n_sims = 25)
#> $stats
#>            [,1]
#> [1,]  0.6463110
#> [2,] -0.5190378
#> [3,] -1.7237974
#> 
#> $sims
#>            tau sim                cat
#> 1  -0.94906994   1 No Drift, No Trend
#> 2  -0.33034201   1    Drift, No Trend
#> 3  -0.76098143   1    Drift and Trend
#> 4   0.72867336   2 No Drift, No Trend
#> 5  -3.89447939   2    Drift, No Trend
#> 6  -0.35265411   2    Drift and Trend
#> 7  -0.29858072   3 No Drift, No Trend
#> 8  -1.01328712   3    Drift, No Trend
#> 9  -1.38086918   3    Drift and Trend
#> 10  0.44315455   4 No Drift, No Trend
#> 11 -0.38454231   4    Drift, No Trend
#> 12 -0.38534101   4    Drift and Trend
#> 13 -1.26057613   5 No Drift, No Trend
#> 14 -1.13544204   5    Drift, No Trend
#> 15 -3.86627995   5    Drift and Trend
#> 16  0.10448858   6 No Drift, No Trend
#> 17 -2.29154139   6    Drift, No Trend
#> 18 -0.45861925   6    Drift and Trend
#> 19  0.90482863   7 No Drift, No Trend
#> 20 -1.52561559   7    Drift, No Trend
#> 21 -0.62745623   7    Drift and Trend
#> 22 -2.83600206   8 No Drift, No Trend
#> 23 -1.18131175   8    Drift, No Trend
#> 24 -0.82316803   8    Drift and Trend
#> 25 -0.15776487   9 No Drift, No Trend
#> 26 -4.12466156   9    Drift, No Trend
#> 27 -1.85473189   9    Drift and Trend
#> 28  0.22520334  10 No Drift, No Trend
#> 29 -2.19042023  10    Drift, No Trend
#> 30  1.41388036  10    Drift and Trend
#> 31  1.06670537  11 No Drift, No Trend
#> 32  0.46118462  11    Drift, No Trend
#> 33 -2.76170987  11    Drift and Trend
#> 34 -0.16015121  12 No Drift, No Trend
#> 35 -3.27307304  12    Drift, No Trend
#> 36 -0.83819276  12    Drift and Trend
#> 37 -0.52800685  13 No Drift, No Trend
#> 38 -0.67430061  13    Drift, No Trend
#> 39  0.27685255  13    Drift and Trend
#> 40  0.16936470  14 No Drift, No Trend
#> 41 -0.72456618  14    Drift, No Trend
#> 42 -4.05140703  14    Drift and Trend
#> 43 -0.05601073  15 No Drift, No Trend
#> 44 -1.46161703  15    Drift, No Trend
#> 45 -3.50235672  15    Drift and Trend
#> 46 -0.42902252  16 No Drift, No Trend
#> 47  0.48771297  16    Drift, No Trend
#> 48 -2.69787501  16    Drift and Trend
#> 49 -1.08581304  17 No Drift, No Trend
#> 50 -0.46880382  17    Drift, No Trend
#> 51  0.20583467  17    Drift and Trend
#> 52  0.49828249  18 No Drift, No Trend
#> 53 -0.99922250  18    Drift, No Trend
#> 54 -4.78599842  18    Drift and Trend
#> 55 -0.99069294  19 No Drift, No Trend
#> 56 -2.04064961  19    Drift, No Trend
#> 57 -1.84882080  19    Drift and Trend
#> 58 -0.15860283  20 No Drift, No Trend
#> 59 -2.14101534  20    Drift, No Trend
#> 60 -1.43000909  20    Drift and Trend
#> 61 -0.97110553  21 No Drift, No Trend
#> 62 -3.72745981  21    Drift, No Trend
#> 63 -0.62934724  21    Drift and Trend
#> 64  0.27594076  22 No Drift, No Trend
#> 65 -2.00897926  22    Drift, No Trend
#> 66 -0.42376049  22    Drift and Trend
#> 67  0.03951816  23 No Drift, No Trend
#> 68 -3.35811779  23    Drift, No Trend
#> 69 -2.39444396  23    Drift and Trend
#> 70 -0.85301050  24 No Drift, No Trend
#> 71 -3.93138856  24    Drift, No Trend
#> 72 -0.48505356  24    Drift and Trend
#> 73 -0.13377280  25 No Drift, No Trend
#> 74 -2.84964556  25    Drift, No Trend
#> 75 -2.78939060  25    Drift and Trend
#> 
#> $attributes
#>   lags       sim_hyp n_sims  n test
#> 1    8 nonstationary     25 25  adf
#> 
#> attr(,"class")
#> [1] "sadf_test"
sadf_test(x, n_sims = 25)
#> $stats
#>            [,1]
#> [1,]  0.2847004
#> [2,] -0.8406674
#> [3,] -2.8309772
#> 
#> $sims
#>            tau sim                cat
#> 1   0.20218408   1 No Drift, No Trend
#> 2  -4.02971750   1    Drift, No Trend
#> 3  -1.26080028   1    Drift and Trend
#> 4  -0.71952142   2 No Drift, No Trend
#> 5  -1.08039849   2    Drift, No Trend
#> 6  -1.01653136   2    Drift and Trend
#> 7  -0.47472057   3 No Drift, No Trend
#> 8  -1.25669039   3    Drift, No Trend
#> 9   0.43549610   3    Drift and Trend
#> 10  2.66379063   4 No Drift, No Trend
#> 11  0.33823384   4    Drift, No Trend
#> 12 -0.81920202   4    Drift and Trend
#> 13  0.08564310   5 No Drift, No Trend
#> 14  0.01071294   5    Drift, No Trend
#> 15 -0.90406531   5    Drift and Trend
#> 16  0.63017118   6 No Drift, No Trend
#> 17 -2.16235385   6    Drift, No Trend
#> 18 -1.58818843   6    Drift and Trend
#> 19  0.09119277   7 No Drift, No Trend
#> 20 -1.80831367   7    Drift, No Trend
#> 21 -1.31477488   7    Drift and Trend
#> 22  0.28439951   8 No Drift, No Trend
#> 23 -1.19208035   8    Drift, No Trend
#> 24 -2.22864618   8    Drift and Trend
#> 25  0.22024121   9 No Drift, No Trend
#> 26 -3.64567224   9    Drift, No Trend
#> 27 -0.22446225   9    Drift and Trend
#> 28  0.18690254  10 No Drift, No Trend
#> 29 -1.77427723  10    Drift, No Trend
#> 30 -0.77391766  10    Drift and Trend
#> 31 -0.35703534  11 No Drift, No Trend
#> 32  0.17671669  11    Drift, No Trend
#> 33 -1.85486309  11    Drift and Trend
#> 34  0.93593969  12 No Drift, No Trend
#> 35 -0.38178989  12    Drift, No Trend
#> 36 -0.72111976  12    Drift and Trend
#> 37 -0.36750799  13 No Drift, No Trend
#> 38 -1.75995185  13    Drift, No Trend
#> 39 -1.87517457  13    Drift and Trend
#> 40 -0.89938904  14 No Drift, No Trend
#> 41 -0.92023272  14    Drift, No Trend
#> 42 -1.93505841  14    Drift and Trend
#> 43 -0.13489836  15 No Drift, No Trend
#> 44 -4.63837867  15    Drift, No Trend
#> 45 -0.29622973  15    Drift and Trend
#> 46  1.12215067  16 No Drift, No Trend
#> 47 -1.08506638  16    Drift, No Trend
#> 48 -6.01468058  16    Drift and Trend
#> 49  0.09021136  17 No Drift, No Trend
#> 50 -1.08778294  17    Drift, No Trend
#> 51 -1.73025103  17    Drift and Trend
#> 52 -0.79663681  18 No Drift, No Trend
#> 53 -2.31382426  18    Drift, No Trend
#> 54 -1.41589019  18    Drift and Trend
#> 55 -1.23049383  19 No Drift, No Trend
#> 56 -2.74146268  19    Drift, No Trend
#> 57  0.12647729  19    Drift and Trend
#> 58 -0.47822220  20 No Drift, No Trend
#> 59 -1.85699615  20    Drift, No Trend
#> 60 -2.57817404  20    Drift and Trend
#> 61 -0.38479252  21 No Drift, No Trend
#> 62  1.50943399  21    Drift, No Trend
#> 63 -1.83440299  21    Drift and Trend
#> 64  2.48808997  22 No Drift, No Trend
#> 65  1.76354398  22    Drift, No Trend
#> 66 -0.56209513  22    Drift and Trend
#> 67 -3.56282164  23 No Drift, No Trend
#> 68 -3.70064239  23    Drift, No Trend
#> 69 -0.32378174  23    Drift and Trend
#> 70 -0.52242206  24 No Drift, No Trend
#> 71 -1.12943011  24    Drift, No Trend
#> 72 -0.12503124  24    Drift and Trend
#> 73  0.52749081  25 No Drift, No Trend
#> 74 -1.33411515  25    Drift, No Trend
#> 75 -4.14348084  25    Drift and Trend
#> 
#> $attributes
#>   lags       sim_hyp n_sims  n test
#> 1    8 nonstationary     25 25  adf
#> 
#> attr(,"class")
#> [1] "sadf_test"
sadf_test(y, n_sims = 25)
#> Error in is.finite(x): default method not implemented for type 'list'