Simulate a Phillips-Perron Test to Assess Unit Root in a Time Series

spp_test() provides a simulation approach to assessing unit root in a time series by way of the Phillips-Perron test. It takes a vector and performs three Phillips-Perron tests (no drift, no trend; drift, no trend; drift and trend) and calculates both rho and tau statistics as one normally would. Rather than interpolate or approximate a p-value, it simulates some user-specified number of Phillips-Perron tests of either a known, non-stationary time series or a known, white-noise time series matching the length of the time series the user provides. This allows the user to make assessments of non-stationarity or stationarity by way of simulation rather than approximation from received critical values by way of books or tables some years out of date.

Usage

spp_test(x, lag_short = TRUE, n_sims = 1000, sim_hyp = "nonstationary")

Arguments

x

a vector

lag_short

logical, defaults to TRUE. If TRUE, the "short-term" lag is used for the Phillips-Perron test. If FALSE, the "long-term" lag is used.

n_sims

the number of simulations for calculating an interval or distribution of test statistics for assessing stationarity or non-stationarity. Defaults to 1,000.

sim_hyp

can be either "stationary" or "nonstationary". If "stationary", the function runs Phillips-Perron tests on simulated stationary (pure white noise) data. This allows the user to assess compatibility/plausibility of the test statistic against a distribution of test statistics that are known to be pure white noise (in expectation). If "nonstationary" (default), the function generates three different data sets of a pure random walk, a random walk with a drift, and a random walk with a drift and trend. It then runs Phillips-Perron tests on all those. This allows the user to assess the compatibility/plausibility of their test statistics with data that are known to be nonstationary in some form.

Value

spp_test() returns a list of length 3. The first element in the list is a matrix of rho statistics and tau statistics calculated by the Phillips-Perron test. The second element is a data frame of the simulated rho and tau statistics of either a known white-noise time series or three different non-stationary time series (pure random walk, random walk with drift, random walk with drift and trend). The third element is some attributes about the procedure for post-processing.

Details

Some knowledge of Augmented Dickey-Fuller and the Phillips-Perron procedure is assumed here. Generally, the Phillips-Perron test purports to build on the Augmented Dickey-Fuller procedure through two primary means. The first is relaxing the need to specify or assume lag structures ad hoc or ex ante. Only a short-term lag or long-term lag are necessary. The second is that its robust to various forms of heteroskedasticity in the error term.

The short-term and long-term lags follow the convention introduced in the Phillips-Perron test. The short-term lag uses the default number of Newey-West lags, defined as the floor of 4*(n/100)^.25 where n is the length of the time series. The long-term lag substitutes 4 for 12 in this equation.

This function specifies three different types of tests: 1) no drift, no trend, 2) drift, no trend, and 3) drift and trend. In the language of the lm() function, the first is lm(y ~ ly - 1) where y is the value of y and ly is its first-order lag. The second test is lm(y ~ ly), intuitively suggesting the y-intercept in this equation is the "drift". The third would be lm(y ~ ly + t) with t being a simple integer that increases by 1 for each observation (i.e. a time-trend).

There are two types of statistics in the Phillips-Perron test: rho and tau. Of the two, tau is the more intuitive statistic and compares favorably to its corollary statistic in the Augmented Dickey-Fuller test. It's why you'll typically see tau reported as the statistic of interest in other implementations. rho has its utility for more advanced diagnostics, though. Both are calculated in this function, though tau is the default statistic.

None of this is meant to discourage the use of Fuller (1976) or its various reproductions for the sake of diagnosing stationarity or non-stationary, and I will confess their expertise on these matters outpaces mine. Consider the justification for this function to be largely philosophical and/or experimental. Why not simulate it? It's not like time or computing power are huge issues anymore.

This is always awkwardly stated, but it's a good reminder that the classic Dickey-Fuller statistics are mostly intended to come back negative. That's not always the case, to be clear, but it is the intended case. You assess the statistic by "how negative" it is. Stationary time series will produce test statistics more negative ("smaller") than those produced by non-stationary time series. In a way, this makes the hypotheses implicitly one-tailed (to use that language).

This function removes missing values from the vector before calculating test statistics.

Author

Steven V. Miller

Examples

a <- rnorm(25) # white noise
b <- cumsum(a) # random walk

spp_test(a, n_sims = 25)
#> $stats
#>           [,1]      [,2]
#> [1,] -27.33738 -4.989217
#> [2,] -27.37505 -4.891018
#> [3,] -30.05658 -5.717812
#> 
#> $sims
#>           z_rho       z_tau sim                cat
#> 1   -0.76193257 -0.45327841   1 No Drift, No Trend
#> 2   -4.19786958 -1.43229335   1    Drift, No Trend
#> 3  -16.92749527 -3.38637319   1    Drift and Trend
#> 4   -0.03709477 -0.02060497   2 No Drift, No Trend
#> 5   -0.95804358 -0.52558260   2    Drift, No Trend
#> 6  -14.87705704 -3.79128833   2    Drift and Trend
#> 7   -1.37787407 -0.86663638   3 No Drift, No Trend
#> 8   -3.39878336 -1.74250861   3    Drift, No Trend
#> 9  -11.62766673 -2.63054568   3    Drift and Trend
#> 10 -17.23448127 -3.56203792   4 No Drift, No Trend
#> 11 -17.45321487 -3.54269783   4    Drift, No Trend
#> 12 -15.44004124 -3.29792690   4    Drift and Trend
#> 13  -3.88791078 -1.36575230   5 No Drift, No Trend
#> 14  -6.19198839 -1.77666418   5    Drift, No Trend
#> 15 -19.95809251 -3.63125323   5    Drift and Trend
#> 16  -6.57373337 -2.00113088   6 No Drift, No Trend
#> 17 -13.07924414 -3.29507684   6    Drift, No Trend
#> 18 -12.79831116 -2.69800929   6    Drift and Trend
#> 19  -2.27829636 -0.97863778   7 No Drift, No Trend
#> 20  -2.30943862 -0.91584312   7    Drift, No Trend
#> 21 -12.50690211 -2.70538984   7    Drift and Trend
#> 22  -1.98438449 -0.99792929   8 No Drift, No Trend
#> 23  -4.02111935 -1.52982436   8    Drift, No Trend
#> 24  -3.33786186 -1.40558542   8    Drift and Trend
#> 25   0.32696041  0.40754845   9 No Drift, No Trend
#> 26  -6.15643055 -2.95459617   9    Drift, No Trend
#> 27  -8.23996855 -2.12235602   9    Drift and Trend
#> 28   2.07479117  1.68218939  10 No Drift, No Trend
#> 29   0.85353199  0.47779642  10    Drift, No Trend
#> 30  -7.02525218 -1.87373475  10    Drift and Trend
#> 31  -0.60163414 -0.51703950  11 No Drift, No Trend
#> 32  -7.21595570 -2.40161631  11    Drift, No Trend
#> 33 -10.18649048 -2.36042617  11    Drift and Trend
#> 34  -0.10795436 -0.14020805  12 No Drift, No Trend
#> 35  -4.96910176 -2.34288923  12    Drift, No Trend
#> 36  -8.13516597 -1.88652973  12    Drift and Trend
#> 37  -0.97140889 -0.51678599  13 No Drift, No Trend
#> 38  -1.84113106 -0.90578220  13    Drift, No Trend
#> 39  -9.47779291 -2.67605666  13    Drift and Trend
#> 40   1.14225668  1.16131076  14 No Drift, No Trend
#> 41  -0.51306027 -0.39092093  14    Drift, No Trend
#> 42 -13.69938740 -2.94693644  14    Drift and Trend
#> 43   0.56642114  0.66820339  15 No Drift, No Trend
#> 44  -1.37051355 -0.98129107  15    Drift, No Trend
#> 45  -9.53595607 -2.31501193  15    Drift and Trend
#> 46  -1.11543747 -0.39157246  16 No Drift, No Trend
#> 47  -2.76878519 -0.88046974  16    Drift, No Trend
#> 48  -1.41748297 -0.69184925  16    Drift and Trend
#> 49   0.34171197  0.21806951  17 No Drift, No Trend
#> 50  -2.91480292 -1.27407207  17    Drift, No Trend
#> 51 -12.21750862 -2.60516948  17    Drift and Trend
#> 52  -4.79367109 -1.64556820  18 No Drift, No Trend
#> 53  -8.39407877 -2.09856462  18    Drift, No Trend
#> 54  -5.83136029 -1.46300970  18    Drift and Trend
#> 55  -0.33462622 -0.14857752  19 No Drift, No Trend
#> 56  -3.01290406 -0.96408476  19    Drift, No Trend
#> 57 -18.17988709 -3.60334190  19    Drift and Trend
#> 58  -3.95705239 -1.45471836  20 No Drift, No Trend
#> 59  -9.22777593 -2.33390624  20    Drift, No Trend
#> 60  -8.54042702 -1.92780252  20    Drift and Trend
#> 61   1.62222738  0.68177848  21 No Drift, No Trend
#> 62  -0.67901539 -0.19997935  21    Drift, No Trend
#> 63 -11.49376380 -2.51094964  21    Drift and Trend
#> 64  -0.31056639 -0.25060570  22 No Drift, No Trend
#> 65  -9.26526902 -2.64229630  22    Drift, No Trend
#> 66 -12.49596265 -2.60571792  22    Drift and Trend
#> 67  -3.05659158 -1.15453427  23 No Drift, No Trend
#> 68  -9.22341380 -2.24546132  23    Drift, No Trend
#> 69  -9.49030268 -2.55404527  23    Drift and Trend
#> 70  -0.19733544 -0.16700540  24 No Drift, No Trend
#> 71  -8.40317911 -3.62216092  24    Drift, No Trend
#> 72  -7.59464351 -1.94056431  24    Drift and Trend
#> 73  -0.05773743 -0.06019397  25 No Drift, No Trend
#> 74  -4.46315272 -1.80966837  25    Drift, No Trend
#> 75 -15.34784472 -2.52801412  25    Drift and Trend
#> 
#> $attributes
#>   lags       sim_hyp n_sims  n test
#> 1    2 nonstationary     25 25   pp
#> 
#> attr(,"class")
#> [1] "spp_test"
spp_test(b, n_sims = 25)
#> $stats
#>           [,1]       [,2]
#> [1,] -1.470509 -0.8571405
#> [2,] -4.527471 -1.5679908
#> [3,] -6.987093 -3.9366364
#> 
#> $sims
#>           z_rho       z_tau sim                cat
#> 1    1.50106127  1.65796630   1 No Drift, No Trend
#> 2    0.06069731  0.06760178   1    Drift, No Trend
#> 3   -6.45369426 -1.83455954   1    Drift and Trend
#> 4   -1.36328906 -0.78803672   2 No Drift, No Trend
#> 5   -6.60045063 -2.29527892   2    Drift, No Trend
#> 6  -20.12039405 -4.19504643   2    Drift and Trend
#> 7    0.10059888  0.12157039   3 No Drift, No Trend
#> 8   -2.79698907 -1.83115075   3    Drift, No Trend
#> 9   -7.41545031 -1.96809228   3    Drift and Trend
#> 10  -0.11432083 -0.11821534   4 No Drift, No Trend
#> 11  -8.26663457 -3.35106944   4    Drift, No Trend
#> 12  -6.97313980 -1.72468515   4    Drift and Trend
#> 13 -17.71511647 -3.37941727   5 No Drift, No Trend
#> 14 -17.83825513 -3.27595539   5    Drift, No Trend
#> 15  -5.21792843 -1.61086005   5    Drift and Trend
#> 16  -0.32064934 -0.12943261   6 No Drift, No Trend
#> 17  -2.98777936 -0.99120421   6    Drift, No Trend
#> 18 -10.05075163 -2.21968827   6    Drift and Trend
#> 19  -1.81276230 -0.91414779   7 No Drift, No Trend
#> 20  -9.45500826 -2.25186412   7    Drift, No Trend
#> 21  -3.18607854 -0.78120055   7    Drift and Trend
#> 22  -1.67921506 -0.90301128   8 No Drift, No Trend
#> 23  -3.37910965 -1.25388027   8    Drift, No Trend
#> 24  -8.80146878 -2.29613094   8    Drift and Trend
#> 25   0.42076845  0.42723550   9 No Drift, No Trend
#> 26  -1.42472115 -1.23729407   9    Drift, No Trend
#> 27  -8.48147824 -2.18899721   9    Drift and Trend
#> 28  -2.37959126 -0.90280044  10 No Drift, No Trend
#> 29  -4.07503525 -1.40397762  10    Drift, No Trend
#> 30 -10.84278775 -1.86023973  10    Drift and Trend
#> 31   0.03107812  0.03472149  11 No Drift, No Trend
#> 32  -5.47015056 -2.10090235  11    Drift, No Trend
#> 33  -6.13002199 -2.05217967  11    Drift and Trend
#> 34   0.78127701  0.98139883  12 No Drift, No Trend
#> 35  -1.16340681 -0.67723884  12    Drift, No Trend
#> 36 -15.03099452 -3.14454130  12    Drift and Trend
#> 37  -0.02679842 -0.02238720  13 No Drift, No Trend
#> 38  -5.07755615 -2.00247436  13    Drift, No Trend
#> 39  -9.23910807 -2.09528824  13    Drift and Trend
#> 40   0.39043561  0.30403986  14 No Drift, No Trend
#> 41  -4.32618645 -2.14420871  14    Drift, No Trend
#> 42 -14.84048430 -3.05601216  14    Drift and Trend
#> 43  -6.56390217 -1.86392098  15 No Drift, No Trend
#> 44  -9.12495390 -2.14461542  15    Drift, No Trend
#> 45  -7.31571612 -1.72177395  15    Drift and Trend
#> 46  -0.04256540 -0.02541170  16 No Drift, No Trend
#> 47  -6.55392636 -1.87355213  16    Drift, No Trend
#> 48  -8.67931691 -2.51868506  16    Drift and Trend
#> 49  -2.89485567 -0.94416253  17 No Drift, No Trend
#> 50  -8.99112293 -2.24321260  17    Drift, No Trend
#> 51  -4.74779709 -1.33929856  17    Drift and Trend
#> 52   1.66768333  2.27666460  18 No Drift, No Trend
#> 53  -0.10721729 -0.08551223  18    Drift, No Trend
#> 54 -12.27841126 -2.64478601  18    Drift and Trend
#> 55  -0.39809497 -0.30444988  19 No Drift, No Trend
#> 56  -3.37642877 -1.24044394  19    Drift, No Trend
#> 57 -11.10494617 -1.70396401  19    Drift and Trend
#> 58  -0.50301990 -0.44796989  20 No Drift, No Trend
#> 59  -4.45746260 -3.20803016  20    Drift, No Trend
#> 60  -5.22790185 -2.16670644  20    Drift and Trend
#> 61  -4.00021020 -1.46501343  21 No Drift, No Trend
#> 62  -9.85891377 -2.44230145  21    Drift, No Trend
#> 63  -5.07009924 -1.28474237  21    Drift and Trend
#> 64  -0.22422235 -0.19947436  22 No Drift, No Trend
#> 65 -11.02873572 -2.93672214  22    Drift, No Trend
#> 66  -9.09464900 -2.92764789  22    Drift and Trend
#> 67  -2.33958877 -0.94207217  23 No Drift, No Trend
#> 68  -7.53865488 -2.25052105  23    Drift, No Trend
#> 69 -13.43317438 -3.89825632  23    Drift and Trend
#> 70   1.63911764  2.53321000  24 No Drift, No Trend
#> 71   0.82351860  0.64067142  24    Drift, No Trend
#> 72  -6.93357539 -1.84081987  24    Drift and Trend
#> 73  -0.87627431 -0.59226483  25 No Drift, No Trend
#> 74  -5.44696598 -1.91720456  25    Drift, No Trend
#> 75  -8.88894944 -2.10364780  25    Drift and Trend
#> 
#> $attributes
#>   lags       sim_hyp n_sims  n test
#> 1    2 nonstationary     25 25   pp
#> 
#> attr(,"class")
#> [1] "spp_test"