Simulate a Phillips-Perron Test to Assess Unit Root in a Time Series

spp_test() provides a simulation approach to assessing unit root in a time series by way of the Phillips-Perron test. It takes a vector and performs three Phillips-Perron tests (no drift, no trend; drift, no trend; drift and trend) and calculates both rho and tau statistics as one normally would. Rather than interpolate or approximate a p-value, it simulates some user-specified number of Phillips-Perron tests of a known, white-noise time series matching the length of the time series the user provides. This allows the user to make assessments of non-stationarity or stationarity by way of simulation rather than approximation from received critical values by way of books or tables some years out of date.

Usage

spp_test(x, lag_short = TRUE, n_sims = 1000, sim_hyp = "nonstationary")

Arguments

x

a vector

lag_short

logical, defaults to TRUE. If TRUE, the "short-term" lag is used for the Phillips-Perron test. If FALSE, the "long-term" lag is used.

n_sims

the number of simulations for calculating an interval or distribution of test statistics for assessing stationarity or non-stationarity. Defaults to 1,000.

sim_hyp

can be either "stationary" or "nonstationary". If "stationary", the function runs Phillips-Perron tests on simulated stationary (pure white noise) data. This allows the user to assess compatibility/plausibility of the test statistic against a distribution of test statistics that are known to be pure white noise (in expectation). If "nonstationary" (default), the function generates three different data sets of a pure random walk, a random walk with a drift, and a random walk with a drift and trend. It then runs Phillips-Perron tests on all those. This allows the user to assess the compatibility/plausibility of their test statistics with data that are known to be nonstationary in some form.

Value

spp_test() returns a list of length 3. The first element in the list is a matrix of rho statistics and tau statistics calculated by the Phillips-Perron test. The second element is a data frame of the simulated rho and tau statistics of either a known white-noise time series or three different non-stationary time series (pure random walk, random walk with drift, random walk with drift and trend). The third element is some attributes about the procedure for post-processing.

Details

Some knowledge of Augmented Dickey-Fuller and the Phillips-Perron procedure is assumed here. Generally, the Phillips-Perron test purports to build on the Augmented Dickey-Fuller procedure through two primary means. The first is relaxing the need to specify or assume lag structures ad hoc or ex ante. Only a short-term lag or long-term lag are necessary. The second is that its robust to various forms of heteroskedasticity in the error term.

The short-term and long-term lags follow the convention introduced in the Phillips-Perron test. The short-term lag uses the default number of Newey-West lags, defined as the floor of 4*(n/100)^.25 where n is the length of the time series. The long-term lag substitutes 4 for 12 in this equation.

This function specifies three different types of tests: 1) no drift, no trend, 2) drift, no trend, and 3) drift and trend. In the language of the lm() function, the first is lm(y ~ ly - 1) where y is the value of y and ly is its first-order lag. The second test is lm(y ~ ly), intuitively suggesting the y-intercept in this equation is the "drift". The third would be lm(y ~ ly + t) with t being a simple integer that increases by 1 for each observation (i.e. a time-trend).

There are two types of statistics in the Phillips-Perron test: rho and tau. Of the two, tau is the more intuitive statistic and compares favorably to its corollary statistic in the Augmented Dickey-Fuller test. It's why you'll typically see tau reported as the statistic of interest in other implementations. rho has its utility for more advanced diagnostics, though. Both are calculated in this function, though tau is the default statistic.

None of this is meant to discourage the use of Fuller (1976) or its various reproductions for the sake of diagnosing stationarity or non-stationary, and I will confess their expertise on these matters outpaces mine. Consider the justification for this function to be largely philosophical and/or experimental. Why not simulate it? It's not like time or computing power are huge issues anymore.

This is always awkwardly stated, but it's a good reminder that the classic Dickey-Fuller statistics are mostly intended to come back negative. That's not always the case, to be clear, but it is the intended case. You assess the statistic by "how negative" it is. Stationary time series will produce test statistics more negative ("smaller") than those produced by non-stationary time series. In a way, this makes the hypotheses implicitly one-tailed (to use that language).

This function removes missing values from the vector before calculating test statistics.

Author

Steven V. Miller

Examples

a <- rnorm(25) # white noise
b <- cumsum(a) # random walk

spp_test(a, n_sims = 25)
#> $stats
#>           [,1]      [,2]
#> [1,] -25.75642 -5.740402
#> [2,] -25.71711 -5.663936
#> [3,] -26.06430 -5.570766
#> 
#> $sims
#>           z_rho       z_tau sim                cat
#> 1    0.70523514  0.58291202   1 No Drift, No Trend
#> 2   -0.91760040 -0.52807748   1    Drift, No Trend
#> 3  -16.02030083 -3.39911261   1    Drift and Trend
#> 4   -0.81568488 -0.41234545   2 No Drift, No Trend
#> 5   -2.53584270 -0.97112505   2    Drift, No Trend
#> 6  -10.00355164 -2.56024506   2    Drift and Trend
#> 7    0.03280584  0.02346779   3 No Drift, No Trend
#> 8   -1.60725846 -0.88158382   3    Drift, No Trend
#> 9  -13.27461508 -3.17073333   3    Drift and Trend
#> 10  -0.74982570 -0.49793857   4 No Drift, No Trend
#> 11  -4.38169745 -1.74739348   4    Drift, No Trend
#> 12  -4.57963082 -1.69414327   4    Drift and Trend
#> 13  -2.37784103 -1.00317968   5 No Drift, No Trend
#> 14  -2.80686991 -1.12096913   5    Drift, No Trend
#> 15  -3.39161415 -1.14608064   5    Drift and Trend
#> 16  -4.43408736 -1.51974624   6 No Drift, No Trend
#> 17  -5.59164520 -1.66603651   6    Drift, No Trend
#> 18  -1.47771831 -0.61865657   6    Drift and Trend
#> 19  -0.58276672 -0.44053891   7 No Drift, No Trend
#> 20  -5.43927853 -1.83178975   7    Drift, No Trend
#> 21  -5.87887777 -1.22443743   7    Drift and Trend
#> 22  -0.35105317 -0.35285881   8 No Drift, No Trend
#> 23  -4.21835094 -1.96307817   8    Drift, No Trend
#> 24 -10.98435387 -2.29047934   8    Drift and Trend
#> 25  -1.90368294 -0.98204077   9 No Drift, No Trend
#> 26  -4.66235449 -1.57651408   9    Drift, No Trend
#> 27 -11.77541257 -2.58144848   9    Drift and Trend
#> 28   0.70123491  0.50290010  10 No Drift, No Trend
#> 29  -1.43064961 -0.68977863  10    Drift, No Trend
#> 30 -12.23373187 -2.40892801  10    Drift and Trend
#> 31  -0.17146928 -0.16435724  11 No Drift, No Trend
#> 32  -6.07616435 -1.96714520  11    Drift, No Trend
#> 33  -8.79962630 -1.99644278  11    Drift and Trend
#> 34  -1.50696138 -0.83270019  12 No Drift, No Trend
#> 35  -2.49428420 -0.81001470  12    Drift, No Trend
#> 36  -6.21419893 -2.17261481  12    Drift and Trend
#> 37   0.98997472  1.12226205  13 No Drift, No Trend
#> 38  -2.08174352 -1.13279244  13    Drift, No Trend
#> 39  -2.60042908 -1.02925535  13    Drift and Trend
#> 40  -4.17534385 -1.34494688  14 No Drift, No Trend
#> 41  -5.77367048 -1.65036867  14    Drift, No Trend
#> 42 -13.75942818 -2.97034769  14    Drift and Trend
#> 43  -1.31232504 -0.74253314  15 No Drift, No Trend
#> 44  -3.82409782 -1.43390230  15    Drift, No Trend
#> 45  -6.26231861 -1.50773490  15    Drift and Trend
#> 46  -9.29764057 -2.29124849  16 No Drift, No Trend
#> 47 -15.17298873 -3.19866359  16    Drift, No Trend
#> 48 -15.37972187 -3.11542083  16    Drift and Trend
#> 49   1.09602623  0.43872528  17 No Drift, No Trend
#> 50  -2.15636805 -0.61061157  17    Drift, No Trend
#> 51  -8.13275520 -2.30198863  17    Drift and Trend
#> 52  -1.12424280 -0.78736172  18 No Drift, No Trend
#> 53 -12.90212540 -2.67077742  18    Drift, No Trend
#> 54  -7.40287140 -2.38151664  18    Drift and Trend
#> 55   0.66641963  0.56271210  19 No Drift, No Trend
#> 56  -0.96717513 -0.52950690  19    Drift, No Trend
#> 57  -2.06838145 -1.02376160  19    Drift and Trend
#> 58 -11.12965943 -2.51586182  20 No Drift, No Trend
#> 59 -12.27968355 -2.66053408  20    Drift, No Trend
#> 60  -3.88004731 -1.26535712  20    Drift and Trend
#> 61  -4.58986916 -1.53507487  21 No Drift, No Trend
#> 62  -8.59787739 -2.22391562  21    Drift, No Trend
#> 63  -1.03410324 -0.61002881  21    Drift and Trend
#> 64  -4.50697974 -1.40024103  22 No Drift, No Trend
#> 65  -4.42225025 -1.37820301  22    Drift, No Trend
#> 66  -3.49992115 -1.59656049  22    Drift and Trend
#> 67   0.10428977  0.07213376  23 No Drift, No Trend
#> 68  -2.66277779 -1.13432061  23    Drift, No Trend
#> 69  -6.59985916 -2.70200023  23    Drift and Trend
#> 70  -1.39628281 -0.70422554  24 No Drift, No Trend
#> 71  -6.86237687 -2.05434963  24    Drift, No Trend
#> 72 -15.22679011 -3.67475241  24    Drift and Trend
#> 73  -1.31741495 -0.66012775  25 No Drift, No Trend
#> 74  -1.75790474 -0.85910256  25    Drift, No Trend
#> 75  -4.90397779 -1.64110851  25    Drift and Trend
#> 
#> $attributes
#>   lags       sim_hyp n_sims  n test
#> 1    2 nonstationary     25 25   pp
#> 
#> attr(,"class")
#> [1] "spp_test"
spp_test(b, n_sims = 25)
#> $stats
#>            [,1]      [,2]
#> [1,]  -8.433283 -2.259897
#> [2,] -12.914856 -2.579437
#> [3,] -13.106282 -2.574868
#> 
#> $sims
#>           z_rho       z_tau sim                cat
#> 1   -8.04417352 -1.62168349   1 No Drift, No Trend
#> 2   -8.29666126 -1.67246085   1    Drift, No Trend
#> 3   -4.61071129 -1.32106224   1    Drift and Trend
#> 4    0.75932929  0.93207541   2 No Drift, No Trend
#> 5   -4.46425724 -2.26203540   2    Drift, No Trend
#> 6  -13.53384195 -2.79047220   2    Drift and Trend
#> 7   -5.70347741 -1.73972077   3 No Drift, No Trend
#> 8   -7.98433489 -2.02570977   3    Drift, No Trend
#> 9  -13.43399374 -2.69590184   3    Drift and Trend
#> 10   1.90420223  1.11614049   4 No Drift, No Trend
#> 11   0.82550375  0.44556957   4    Drift, No Trend
#> 12 -13.89184132 -2.72153016   4    Drift and Trend
#> 13   2.05836698  2.16252413   5 No Drift, No Trend
#> 14   0.83614927  0.61413334   5    Drift, No Trend
#> 15  -7.49064048 -2.11180200   5    Drift and Trend
#> 16   1.22164403  1.60089022   6 No Drift, No Trend
#> 17  -1.51171387 -0.88265146   6    Drift, No Trend
#> 18  -8.79578955 -2.06050930   6    Drift and Trend
#> 19  -0.22865523 -0.10790964   7 No Drift, No Trend
#> 20  -5.22276003 -1.30283561   7    Drift, No Trend
#> 21  -8.59359745 -2.12790496   7    Drift and Trend
#> 22   0.76357186  0.99881380   8 No Drift, No Trend
#> 23  -1.43570103 -1.11418515   8    Drift, No Trend
#> 24  -8.31893944 -2.01731461   8    Drift and Trend
#> 25   1.37427784  0.67569299   9 No Drift, No Trend
#> 26  -0.47838528 -0.16678013   9    Drift, No Trend
#> 27  -9.51026293 -2.20768044   9    Drift and Trend
#> 28  -2.33164305 -0.91464602  10 No Drift, No Trend
#> 29  -2.52639917 -0.97450839  10    Drift, No Trend
#> 30  -6.08868918 -1.82676743  10    Drift and Trend
#> 31  -8.58589143 -2.15020175  11 No Drift, No Trend
#> 32  -8.81740839 -2.11233013  11    Drift, No Trend
#> 33  -8.87921030 -2.17261929  11    Drift and Trend
#> 34  -0.74934269 -0.47802179  12 No Drift, No Trend
#> 35  -6.46616426 -2.22879427  12    Drift, No Trend
#> 36 -18.26132863 -3.61314759  12    Drift and Trend
#> 37  -0.58255923 -0.53358962  13 No Drift, No Trend
#> 38  -3.40085733 -1.29426029  13    Drift, No Trend
#> 39  -7.89612187 -1.94031826  13    Drift and Trend
#> 40   0.45859975  0.36700385  14 No Drift, No Trend
#> 41  -1.92342282 -1.09260216  14    Drift, No Trend
#> 42  -1.12861298 -0.83046720  14    Drift and Trend
#> 43   1.80129797  0.95753263  15 No Drift, No Trend
#> 44   0.44732283  0.16767843  15    Drift, No Trend
#> 45  -9.99642088 -2.46510708  15    Drift and Trend
#> 46  -2.04318525 -0.96921349  16 No Drift, No Trend
#> 47  -3.35331980 -1.06021730  16    Drift, No Trend
#> 48  -2.79115781 -1.11744471  16    Drift and Trend
#> 49   1.22131054  0.98746520  17 No Drift, No Trend
#> 50   0.35782518  0.29482423  17    Drift, No Trend
#> 51 -11.02037406 -2.40089495  17    Drift and Trend
#> 52  -6.93118455 -1.68177206  18 No Drift, No Trend
#> 53  -8.24056839 -1.93413518  18    Drift, No Trend
#> 54 -11.66319186 -2.75794315  18    Drift and Trend
#> 55  -0.04992092 -0.04319314  19 No Drift, No Trend
#> 56  -3.56109810 -1.90367361  19    Drift, No Trend
#> 57  -9.91329328 -2.26522215  19    Drift and Trend
#> 58   0.12012911  0.11650759  20 No Drift, No Trend
#> 59  -4.76336447 -1.97274356  20    Drift, No Trend
#> 60  -4.05680267 -0.83856836  20    Drift and Trend
#> 61  -6.69841765 -1.88613264  21 No Drift, No Trend
#> 62 -10.63127819 -2.47770253  21    Drift, No Trend
#> 63 -24.86516223 -5.43562967  21    Drift and Trend
#> 64  -2.42111974 -0.90418202  22 No Drift, No Trend
#> 65  -4.61685699 -1.37794750  22    Drift, No Trend
#> 66 -15.26918336 -3.05171724  22    Drift and Trend
#> 67  -0.92468960 -0.58757330  23 No Drift, No Trend
#> 68  -4.34981838 -1.59025448  23    Drift, No Trend
#> 69  -1.88491980 -0.56927970  23    Drift and Trend
#> 70  -1.35132736 -0.86147574  24 No Drift, No Trend
#> 71  -2.90156943 -1.05101123  24    Drift, No Trend
#> 72  -5.29211479 -1.69715681  24    Drift and Trend
#> 73  -1.45771796 -0.77010710  25 No Drift, No Trend
#> 74  -2.06977908 -1.00700553  25    Drift, No Trend
#> 75  -6.80015873 -1.82153724  25    Drift and Trend
#> 
#> $attributes
#>   lags       sim_hyp n_sims  n test
#> 1    2 nonstationary     25 25   pp
#> 
#> attr(,"class")
#> [1] "spp_test"